pyproject.toml
[project]
name = "test-new-kernel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.66.1",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "missingno>=0.5.2",
    "mlflow>=3.8.1",
    "notebook>=7.5.1",
    "numpy>=1.22,<2.4",
    "optuna>=4.6.0",
    "optuna-integration[xgboost]>=4.6.0",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "statsmodels>=0.14.6",
    "taskipy>=1.14.1",
    "tensorflow-cpu>=2.20.0",
    "xgboost>=3.1.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
norecursedirs = [
    "data", "artifacts", "models", "metrics", "experiments", 
    "__pycache__", ".venv", ".pytest_cache"
]
addopts = "-p no:warnings"

[tool.coverage.run]
source = ["src", "app"]
omit = [
    "data/*",
    "artifacts/*",
    "models/*",
    "metrics/*",
    "experiments/*",
    "*/__pycache__/*",
    ".venv/*",
    ".pytest_cache/*",
    "*.ipynb",
    "tests/*"
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:"
]


[tool.taskipy.tasks]
run = 'fastapi dev app/api/main.py'
test = 'pytest --cov=src --cov=app -vv'
post_test = 'coverage html'
